+++ clients/python/balancing_services/api/default/get_balancing_capacity_procured_volumes.py
@@ -7,7 +7,9 @@ import httpx
 from ... import errors
 from ...client import AuthenticatedClient, Client
 from ...models.area import Area
-from ...models.balancing_capacity_volumes_response import BalancingCapacityVolumesResponse
+from ...models.balancing_capacity_volumes_response import (
+    BalancingCapacityVolumesResponse,
+)
 from ...models.problem import Problem
 from ...models.reserve_type import ReserveType
 from ...types import UNSET, Response
@@ -20,6 +22,7 @@ def _get_kwargs(
     period_end_at: datetime.datetime,
     reserve_type: ReserveType,
 ) -> dict[str, Any]:
+
     params: dict[str, Any] = {}
 
     json_area = area.value