+++ clients/python/balancing_services/models/currency.py
@@ -3,6 +3,7 @@ from enum import Enum
 
 class Currency(str, Enum):
     BGN = "BGN"
+    CHF = "CHF"
     EUR = "EUR"
     HUF = "HUF"
     PLN = "PLN"