+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for CHF (Swiss Franc) currency
+
 ## [1.6.0] - 2026-01-30
 
 ### Added